fix: fix a bug with the arguements passed to ocm cluter mgmt and rename list cluster package
