+++ pkg/cmd/dedicated/dedicated.go
@@ -2,7 +2,7 @@ package dedicated
 
 import (
 	"github.com/redhat-developer/app-services-cli/pkg/cmd/dedicated/deregister"
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/dedicated/listclusters"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/dedicated/list"
 	"github.com/redhat-developer/app-services-cli/pkg/cmd/dedicated/register"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
 	"github.com/spf13/cobra"
@@ -18,7 +18,7 @@ func NewDedicatedCmd(f *factory.Factory) *cobra.Command {
 
 	cmd.AddCommand(
 		register.NewRegisterClusterCommand(f),
-		listclusters.NewListClusterCommand(f),
+		list.NewListClusterCommand(f),
 		deregister.NewDeRegisterClusterCommand(f),
 	)
