+++ pkg/cmd/dedicated/list/list_test.go
@@ -1,4 +1,4 @@
-package listclusters
+package list
 
 import (
 	"testing"
